Fix int vs. size_t in a test output fprintf Fix privacy of position and parent in references to iterator -- HJB

+++ trunk/NearTree/CNearTreeTest.cpp
@@ -2765,5 +2765,5 @@ void testSeparation( void )
     typedef std::vector<int> container;
     container group1, group2;
     t.BelongsToPoints( imin, imax, group1, group2 );
-    fprintf( stdout, "in %d out %d\n", group1.size(), group2.size() );
+    fprintf( stdout, "in %ld out %ld\n", (long int)group1.size(), (long int)group2.size() );
 }
